+++ torch_geometric/datasets/airfrans.py
@@ -65,6 +65,8 @@ class AirfRANS(InMemoryDataset):
             :obj:`torch_geometric.data.Data` object and returns a boolean
             value, indicating whether the data object should be included in the
             final dataset. (default: :obj:`None`)
+        force_reload (bool, optional): Whether to re-process the dataset.
+            (default: :obj:`False`)
 
     **STATS:**
 
@@ -94,6 +96,7 @@ class AirfRANS(InMemoryDataset):
         transform: Optional[Callable] = None,
         pre_transform: Optional[Callable] = None,
         pre_filter: Optional[Callable] = None,
+        force_reload: bool = False,
     ):
         if task not in self.tasks:
             raise ValueError(f"Expected 'task' to be in {self.tasks} "
@@ -102,7 +105,8 @@ class AirfRANS(InMemoryDataset):
         self.task = 'full' if task == 'scarce' and not train else task
         self.split = 'train' if train else 'test'
 
-        super().__init__(root, transform, pre_transform, pre_filter)
+        super().__init__(root, transform, pre_transform, pre_filter,
+                         force_reload=force_reload)
         self.load(self.processed_paths[0])
 
     @property