+++ torch_geometric/datasets/amazon_products.py
@@ -25,6 +25,8 @@ class AmazonProducts(InMemoryDataset):
             an :obj:`torch_geometric.data.Data` object and returns a
             transformed version. The data object will be transformed before
             being saved to disk. (default: :obj:`None`)
+        force_reload (bool, optional): Whether to re-process the dataset.
+            (default: :obj:`False`)
 
     **STATS:**
 
@@ -48,9 +50,15 @@ class AmazonProducts(InMemoryDataset):
     class_map_id = '1LIl4kimLfftj4-7NmValuWyCQE8AaE7P'
     role_id = '1npK9xlmbnjNkV80hK2Q68wTEVOFjnt4K'
 
-    def __init__(self, root: str, transform: Optional[Callable] = None,
-                 pre_transform: Optional[Callable] = None):
-        super().__init__(root, transform, pre_transform)
+    def __init__(
+        self,
+        root: str,
+        transform: Optional[Callable] = None,
+        pre_transform: Optional[Callable] = None,
+        force_reload: bool = False,
+    ):
+        super().__init__(root, transform, pre_transform,
+                         force_reload=force_reload)
         self.load(self.processed_paths[0])
 
     @property