+++ torch_geometric/datasets/pcpnet_dataset.py
@@ -44,6 +44,8 @@ class PCPNetDataset(InMemoryDataset):
             :obj:`torch_geometric.data.Data` object and returns a boolean
             value, indicating whether the data object should be included in the
             final dataset. (default: :obj:`None`)
+        force_reload (bool, optional): Whether to re-process the dataset.
+            (default: :obj:`False`)
     """
 
     url = 'http://geometry.cs.ucl.ac.uk/projects/2018/pcpnet/pclouds.zip'
@@ -80,6 +82,7 @@ class PCPNetDataset(InMemoryDataset):
         transform: Optional[Callable] = None,
         pre_transform: Optional[Callable] = None,
         pre_filter: Optional[Callable] = None,
+        force_reload: bool = False,
     ):
 
         assert split in ['train', 'val', 'test']
@@ -94,7 +97,8 @@ class PCPNetDataset(InMemoryDataset):
         self.category = category
         self.split = split
 
-        super().__init__(root, transform, pre_transform, pre_filter)
+        super().__init__(root, transform, pre_transform, pre_filter,
+                         force_reload=force_reload)
         self.load(self.processed_paths[0])
 
     @property